Moving sources/outline stuff into a "PrimaryPanes" component/dir.
