+++ src/components/PrimaryPanes/Outline.css
